+++ packages/codemods/src/codeMods/__tests__/utilities/importUtils.test.ts
@@ -1,4 +1,4 @@
-import { utilities } from '../../utilities/utilities';
+import { getImportsByPath, repathImport } from '../../utilities';
 import { Project } from 'ts-morph';
 
 const fileName = 'mockImports.tsx';
@@ -15,20 +15,20 @@ describe('Import Utilities test', () => {
 
   it('can find import based on exact string match', () => {
     const file = project.getSourceFileOrThrow(fileName);
-    const tag = utilities.getImportsByPath(file, buttonPath);
+    const tag = getImportsByPath(file, buttonPath);
     expect(tag).toBeTruthy();
     expect(tag![0].getModuleSpecifierValue()).toEqual(buttonPath);
   });
 
   it('only finds exact matches', () => {
     const file = project.getSourceFileOrThrow(fileName);
-    const imp = utilities.getImportsByPath(file, rootPath);
+    const imp = getImportsByPath(file, rootPath);
     expect(imp.length).toEqual(0);
   });
 
   it('can find a single import based on a regex string match', () => {
     const file = project.getSourceFileOrThrow(fileName);
-    const imp = utilities.getImportsByPath(file, /office\-ui\-fabric\-react.+Button/);
+    const imp = getImportsByPath(file, /office\-ui\-fabric\-react.+Button/);
     expect(imp).toBeTruthy();
     expect(imp?.length).toEqual(1);
     expect(imp![0].getModuleSpecifierValue()).toEqual(buttonPath);
@@ -36,7 +36,7 @@ describe('Import Utilities test', () => {
 
   it('can find all imports based on a regex string match', () => {
     const file = project.getSourceFileOrThrow(fileName);
-    const imps = utilities.getImportsByPath(file, /office\-ui\-fabric\-react/);
+    const imps = getImportsByPath(file, /office\-ui\-fabric\-react/);
     expect(imps).toBeTruthy();
     expect(imps?.length).toBeGreaterThan(1);
     imps!.forEach(imp => {
@@ -47,9 +47,9 @@ describe('Import Utilities test', () => {
   it('can replace an import path', () => {
     const replacementString = 'Complete/NewPath';
     const file = project.getSourceFileOrThrow(fileName);
-    const imps = utilities.getImportsByPath(file, /office\-ui\-fabric\-react/);
+    const imps = getImportsByPath(file, /office\-ui\-fabric\-react/);
 
-    utilities.repathImport(imps![0], replacementString);
+    repathImport(imps![0], replacementString);
     expect(
       file.getImportStringLiterals().some(val => {
         return val.getLiteralValue() === replacementString;
@@ -60,9 +60,9 @@ describe('Import Utilities test', () => {
   it('can replace a partial import path', () => {
     const replacementRegex = /office\-ui\-fabric\-react/;
     const file = project.getSourceFileOrThrow(fileName);
-    const imps = utilities.getImportsByPath(file, /office\-ui\-fabric\-react/);
+    const imps = getImportsByPath(file, /office\-ui\-fabric\-react/);
 
-    utilities.repathImport(imps![0], 'NewPath', replacementRegex);
+    repathImport(imps![0], 'NewPath', replacementRegex);
     expect(
       file.getImportStringLiterals().some(val => {
         return val.getLiteralValue() === 'NewPath/lib/Button';
