+++ packages/codemods/src/codeMods/__tests__/utilities/jsxUtilities.test.ts
@@ -1,4 +1,4 @@
-import { utilities } from '../../utilities/utilities';
+import { findJsxTag, renameImport } from '../../utilities';
 import { Project, SourceFile } from 'ts-morph';
 
 const project = new Project();
@@ -7,18 +7,18 @@ const fileName = 'mockFunction.tsx';
 
 describe('JSX Utilities Test', () => {
   it('can find a regular jsx tag', () => {
-    const tag = utilities.findJsxTag(project.getSourceFileOrThrow(fileName), 'JSXFunctionalNormalTag');
+    const tag = findJsxTag(project.getSourceFileOrThrow(fileName), 'JSXFunctionalNormalTag');
     expect(tag.length).toEqual(1);
   });
 
   it('can find a self closing jsx tag', () => {
-    const tag = utilities.findJsxTag(project.getSourceFileOrThrow(fileName), 'JSXFunctionalSelfClosingTag');
+    const tag = findJsxTag(project.getSourceFileOrThrow(fileName), 'JSXFunctionalSelfClosingTag');
     expect(tag.length).toEqual(1);
   });
 
   it('can replace a tag', () => {
     const file = project.getSourceFileOrThrow(fileName);
-    utilities.renameImport(file, 'ToImport', 'Renamed');
+    renameImport(file, 'ToImport', 'Renamed');
     expect(file.getText().indexOf('ToImport')).toBe(-1);
   });
 });
