+++ packages/codemods/src/codeMods/mods/PersonaToAvatar/PersonaToAvatar.mod.ts
@@ -8,7 +8,7 @@ import {
   ts,
   Node,
 } from 'ts-morph';
-import { utilities } from '../../utilities/utilities';
+import { findJsxTag } from '../../utilities';
 import { ICodeMod } from '../../ICodeMod';
 
 const personaPath = 'office-ui-fabric-react/lib/Persona';
@@ -168,14 +168,14 @@ function getBlockContainer(node: Node<ts.Node>) {
 
 export function RenamePrimaryTextProp(file: SourceFile) {
   // Should this fix the naming if the Persona Component has already been renamed to Avatar
-  const elements = utilities.findJsxTag(file, 'Persona');
+  const elements = findJsxTag(file, 'Persona');
   renameProperty(elements, 'primaryText', 'text');
 }
 
 export function RenameRenderCoin(file: SourceFile) {
   // Should this fix the naming if the Persona Component has already been renamed to Avatar
 
-  const elements = utilities.findJsxTag(file, 'Persona');
+  const elements = findJsxTag(file, 'Persona');
   renameProperty(elements, 'onRenderCoin', 'onRenderAvatarCoin');
 }
