--- packages/codemods/src/codeMods/utilities/utilities.ts
@@ -1,9 +1,0 @@
-import { findJsxTag } from './jsx';
-import { renameImport, repathImport, getImportsByPath } from './imports';
-
-export const utilities = {
-  findJsxTag,
-  renameImport,
-  repathImport,
-  getImportsByPath,
-};
